Add tests + Refractor old tests file
